Serialize the storm tests: stop starving the suite's tight timeout budgets
Five distinct pre-existing tests have now blown their [Fact(Timeout)]
budgets across CI legs (Race_LateLoserFaultAfterWinner, TestInitialization,
TestAsyncEnumerable twice -- since deflaked -- UpgradeableLock_ThreadSavety,
TestRaceHandling), all with second-scale budgets written for a smaller
suite. The common cause is this branch's storm tests (parallel
writer/reader floods running for seconds) saturating the 2-4 core runners'
thread pool WHILE those tests race their budgets.

The three storms (actor unprimed chain, lock-engine unprimed chain, actor
concurrent sets/gets) move to dedicated classes in an xunit collection
with DisableParallelization = true: they run sequentially, alongside
nothing, so the rest of the suite keeps the scheduling environment its
budgets were written for. Doc references to the moved tests updated.

235 + 48 tests pass locally, complexity gate clean, Coyote systematic
tests pass.
